refactor(dev-server): extract shared devPluginShared (empty discovery + client-module check)
dev-server/plugin.{server,client} duplicated two things verbatim: the empty
AutoDiscoveredFiles literal handed to configureReactServer, and the JS/TS +
detectClientModule 'is this a use-client source file' IIFE in hotUpdate. Extract
both to devPluginShared.ts (emptyAutoDiscoveredFiles, isClientModuleFile).

Scoped to the genuinely-identical pieces; the broader hotUpdate orchestration
legitimately differs per side (server branches on env + sends the WS event;
client does worker invalidation with an isProcessingHmr guard) and stays in place.

Verified: build clean, 101 entrypoints, test:unit 472, test:dev 37,
test:server 655/4-skip, test:client 182/7-skip. Closes 66g.9.
